refactor: move packet parser to new file
